Replace d.head..d.tail with range of sub-slice
to help compiler eliminate bound checks.

Benchstat results:
```
name                         old time/op    new time/op    delta
Decoder_Base64Append/128-4      507ns ± 2%     462ns ± 2%   -8.92%  (p=0.000 n=15+13)
Decoder_Base64Append/256-4      954ns ± 1%     865ns ± 1%   -9.32%  (p=0.000 n=13+13)
Decoder_Base64Append/512-4     1.86µs ± 1%    1.68µs ± 3%   -9.75%  (p=0.000 n=14+15)
Decoder_Base64Append/1024-4    3.67µs ± 1%    3.28µs ± 1%  -10.68%  (p=0.000 n=15+15)
OTEL_Decode-4                  1.55µs ± 2%    1.48µs ± 2%   -4.60%  (p=0.000 n=15+15)

name                         old speed      new speed      delta
Decoder_Base64Append/128-4    252MB/s ± 2%   277MB/s ± 2%   +9.79%  (p=0.000 n=15+13)
Decoder_Base64Append/256-4    268MB/s ± 1%   296MB/s ± 1%  +10.28%  (p=0.000 n=13+13)
Decoder_Base64Append/512-4    275MB/s ± 1%   305MB/s ± 3%  +10.82%  (p=0.000 n=14+15)
Decoder_Base64Append/1024-4   279MB/s ± 1%   312MB/s ± 1%  +11.95%  (p=0.000 n=15+15)
OTEL_Decode-4                 333MB/s ± 2%   350MB/s ± 2%   +4.83%  (p=0.000 n=15+15)
```